debian: Add rules to create a +dfsg tarball
At the next release, this will require
 * Creating a tag on github like "v3.2.0"
 * Using a Debian version number like "3.2.0+dsfg-1"
but it will get us what we want for a Debian package (a source tarball
that excludes the non-DFSG-free 'winio' packages) without breaking the
mesaflash git repository's ability to build for/on Windows.

Since there are no tags at the moment, I tested this by temporarily
naming jepler/mesaflash in debian/watch, and creating a tag there.
It seems to work.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>